Update v2.5.4
This version fixes the cluster and singleton names in the clusterCompostition.tsv file when a previous nomenclature was provided. As the clusterComposition file is created before the nomenclature file is analyzed, in previous versions of ReporTree, the cluster names in this file (and only this file) did not correspond to the cluster nomenclature provided. This has been fixed from version 2.5.4 on.